Deleted onlyOwner modifier on addMember function
Deleted the onlyOwner modifier on the addMember function for testing purposes. Invalid opCode errors would be thrown otherwise, since the test contract is not the owner of the deployed Congress contract. Can be revised later if errors are fixed in some other way.
Also changed the view of getContractAddress to external.