Added function proposalExists and memberHasVoted
Added variable to struct Proposal, which signals wether or not data has been added to the proposal. Reworked the function memberExists and added two new functions: proposalExists, which checks wether or not a given proposal exists, and memberHasVoted, which checks wether or not a given member has voted on a specific proposal.