Added Shareholder Functionality
Added functionality for shareholders. Each member is now assigned a weight for their vote. This weight must be an unsigned integer (256 bit) and will be added or subtracted from the final score of the proposal, depending on wether their vote was in favor of the proposal or not. Additionally the majority margin and the final score of each proposal now also has to be an unsigned integer (256 bit).